refactor(interface-cli): extract clap argument tree into args.rs
Pull the full clap-derive argument tree (`Cli`, `Command`, the 6
subcommand enums) and the `cli_parse_tests` module out of `main.rs`
into a new `crates/interface-cli/src/args.rs`:

- `Cli` (top-level parser + global flags)
- `Command` (15 variants)
- `MigrateCommand`, `AccountCommand`, `ApiKeysCommand`,
  `OrchestratorCommand`, `WebUiCommand`, `PersonaCommand`, `SkillCommand`
- `cli_parse_tests` module (12 tests)

All enums + the `Cli` struct flip to `pub`. `cmd_persona.rs`,
`cmd_skill.rs`, and `cmd_webui.rs` now import from `crate::args::*`
rather than `crate::*`.

main.rs shrinks 1509 → 1002 lines (-507). cargo test -p assistant-cli
reports the same 59 lib tests passing; make lint is clean.

main.rs is now mostly the `main()` REPL function (~960 lines) plus a
handful of top-level imports and the `home_agent_root` helper.
